+++ src/components/dashboardComponents/DashboardOld.js
@@ -1,0 +1,10 @@
+import React, { useState, useEffect } from 'react';
+import Template from './Dashboard';
+
+const Dashboard = () => {
+  return (
+    <Template/>
+  );
+};
+
+export default Dashboard;
